requirements.txt
streamlit
torch
torchvision
torchaudio
transformers
git+https://github.com/openai/CLIP.git
pinecone
pillow
requests
numpy